Modify the plane_segmentation node, now we can test if the around area of the defect fit the feature of a plane. This function provides us with a way to judge if we need to do the B-spline fitting for the defect area or we can just use a plane to work on.
Add a node named area_get in which we can get an area around a specific position. This function can be used to get the area in which a small defect located at or the around area of a large defect

Add a node visualizer, I want to use it to do all the visualizations rather than do them seperately in each node. But it's not finished yet.